+++ examples/redis/main.py
@@ -1,6 +1,10 @@
-import aioredis
+from datetime import date, datetime
+import time
+
+import redis.asyncio as redis
 import uvicorn
 from fastapi import FastAPI
+from redis.asyncio.connection import ConnectionPool
 from starlette.requests import Request
 from starlette.responses import Response
 
@@ -31,10 +35,30 @@ async def clear():
     return await FastAPICache.clear(namespace="test")
 
 
+@app.get("/date")
+@cache(namespace="test", expire=20)
+async def get_data(request: Request, response: Response):
+    return date.today()
+
+
+@app.get("/blocking")
+@cache(namespace="test", expire=20)
+def blocking(request: Request, response: Response):
+    time.sleep(5)
+    return dict(ret=get_ret())
+
+
+@app.get("/datetime")
+@cache(namespace="test", expire=20)
+async def get_datetime(request: Request, response: Response):
+    return datetime.now()
+
+
 @app.on_event("startup")
 async def startup():
-    redis = aioredis.from_url(url="redis://localhost")
-    FastAPICache.init(RedisBackend(redis), prefix="fastapi-cache")
+    pool = ConnectionPool.from_url(url="redis://localhost")
+    r = redis.Redis(connection_pool=pool)
+    FastAPICache.init(RedisBackend(r), prefix="fastapi-cache")
 
 
 if __name__ == "__main__":